requirements.txt
PyYAML
sympy
transformers
accelerate
sentencepiece